Move functions and other few changes
