+++ Pose2Sim/Demo_Batch/Trial_1/Config.toml
@@ -19,7 +19,7 @@
 
 # [project]
 # multi_person = false # true for trials with multiple participants. If false, only the main person in scene is analyzed (and it run much faster). 
-# participant_height = 1.72 # m # float if single person, list of float if multi-person (same order as the Static trials) # Only used for marker augmentation
+# participant_height = 'auto' # m # Set it to 'auto' for automatic compute mode # float if single person, list of float if multi-person (same order as the Static trials) # Only used for marker augmentation
 # participant_mass = 70.0 # kg # Only used for marker augmentation and scaling
 
 # frame_rate = 'auto' # fps # int or 'auto'. If 'auto', finds from video (or defaults to 60 fps if you work with images) 
@@ -178,7 +178,10 @@
 # ##        "RAnkle", "LAnkle", "RHeel", "LHeel", "RSmallToe", "LSmallToe",
 # ##        "RBigToe", "LBigToe", "RElbow", "LElbow", "RWrist", "LWrist"]
 # make_c3d = false # save triangulated data in c3d format in addition to trc
-
+# fastest_frames_to_remove_percent = 0.1 # Frames with high speed are considered as outliers
+# close_to_zero_speed_m = 0.2 # Sum for all keypoints: about 50 px/frame or 0.2 m/frame
+# large_hip_knee_angles = 45 # Hip and knee angles below this value are considered as imprecise
+# trimmed_extrema_percent = 0.5 # Proportion of the most extreme segment values to remove before calculating their mean)
 
 # [kinematics]
 # use_augmentation = true  # true or false (lowercase) # Set to true if you want to use the model with augmented markers
@@ -186,10 +189,7 @@
 
 # remove_individual_scaling_setup = true # true or false (lowercase) # If true, the individual scaling setup files are removed to avoid cluttering
 # remove_individual_IK_setup = true # true or false (lowercase) # If true, the individual IK setup files are removed to avoid cluttering
-# fastest_frames_to_remove_percent = 0.1 # Frames with high speed are considered as outliers
-# close_to_zero_speed_m = 0.2 # Sum for all keypoints: about 50 px/frame or 0.2 m/frame
-# large_hip_knee_angles = 45 # Hip and knee angles below this value are considered as imprecise
-# trimmed_extrema_percent = 0.5 # Proportion of the most extreme segment values to remove before calculating their mean)
+
 
 
 # # CUSTOM skeleton, if you trained your own model from DeepLabCut or MMPose for example.
