+++ Pose2Sim/Demo_MultiPerson/Config.toml
@@ -19,7 +19,7 @@
 
 [project]
 multi_person = true # true for trials with multiple participants. If false, only the main person in scene is analyzed (and it run much faster). 
-participant_height = [1.72, 1.40] # m # Set it to 'auto' for automatic compute mode # float if single person, list of float if multi-person (same order as the Static trials) # Only used for marker augmentation
+participant_height = 'auto' # m # Set it to 'auto' for automatic compute mode # float if single person, list of float if multi-person (same order as the Static trials) # Only used for marker augmentation
 participant_mass = [70.0, 63.5] # kg # Only used for marker augmentation and scaling
 
 frame_rate = 'auto' # fps # int or 'auto'. If 'auto', finds from video (or defaults to 60 fps if you work with images)
